Update Texture.cs
The original version caused a design time error because the parameter of one method was of type string, but that same method returned Texture with a parameter of type int. Also, I'm relatively certain that it is not a thing to have a method signature in the signature of a constructor...I could be mistaken, though.